Added a <Link> component
All links now use this component instead of Nextjs' one. This will determine if the `href` prop is an internal or external link, then decide what to do. For external links, `<a href={href} rel="external">...</a>` is used (note the use of `rel` instead of `className`; `css/global.css` has been updated for this change). For internal links, `/instruction/...` and the others are handled differently; `/instruction/` (where `` is not `help`) use the Nextjs <Link> with `href="/instruction/[slug]" as={href}` props. All other internal links (including `/instruction/help`) use the Nextjs <Link> as normal