Added a "Bit Encoding" section to instruction pages
This section lists the actual bit representation of an instruction.
The information is pulled from Appendix B of Volume 2 of the Intel
x86 SDM. The source is listed in `bitEncoding[].src`. If there
isn't one (for example, ADCX and ADOX), the value `none` is used.

Any patterns I pick out while building these data files will be
noted in `/data/instructionNotes.txt`.

Currently, only non-AVX forms are listed, but this will change. For
example, ADDPD only lists the LEGACY encoding, not the VEX or EVEX
forms.

Also, the `processStringToJsx` function now recognizes `\bits` and
`\bitRef`. Currently, `\bits` is a stub copy of `\c`, but this may
change.

Finally, a crude `refs` section of the YAML files was added. It is
currently unused, but this will change.