Work around a Next.js issue
The issue is that `<Link href="#headingX">...</Link>` turns into
`<a href="{page}#headingX">...</a>`. Normally, this wouldn't be
that big of an issue, but instruction page TOC links are turning
into this: `<a href="/instruction/[slug]#headingX">...</a>` which
fails.

The (temporary?) solution to this is to use a plain <a> element
instead of the "Next way" of `<Link href={href}><a>...</a></Link>`.